backend/requirements.txt
django==3.2.20
djangorestframework==3.14.0
django-filter==23.2
djoser==2.1.0
Pillow==9.1.0
urllib3==1.26.10